requirements.txt
pygsheets
pandas
gspread
gspread-formatting
google-api-python-client
oauth2client
python-dotenv
